Move remaining legacy parts to legacy module
